requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==3.2.19
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
pip==23.1.2
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
tweepy
python-dotenv